pyproject.toml
[project]
name = "gymnasium-solver-env"
version = "0.1.0"
description = "Environment for gymnasium-solver (uv/jupyter kernel)"
requires-python = ">=3.12"
dependencies = [
    "gymnasium[classic-control,atari,box2d,mujoco,other]",
    "gymnasium-robotics",
    "ale-py",
    "minari[hdf5]",
    "torch",
    "numpy",
    "wandb",
    "wandb-workspaces",
    # Pin OCAtari to avoid broken sdists (e.g., 1.1.5 missing requirements.txt)
    "ocatari>=2.2.1",
    "pytorch-lightning",
    "ipykernel",
    "jupyter",
    "torchvision",
    "pytest",
    "pyyaml",
    "ruamel.yaml",
    "watchdog",
    "setuptools<80",
    "huggingface_hub>=0.22.0",
    "gradio>=4.0.0",
    "vizdoom",
    "mcp",
    "python-dotenv",
    "modal>=0.64.0"
]

[project.optional-dependencies]
retro = [
    # NOTE: stable-retro 0.9.5 is BROKEN on M1 Mac (arm64 wheel contains x86_64 binary)
    # Building from source also fails. Use with Rosetta or wait for upstream fix.
    "stable-retro>=0.9.5",
]

[tool.pytest.ini_options]
# Test discovery patterns
testpaths = ["tests", "learners", "utils"]
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Add the current directory to the Python path so imports work correctly
addopts = [
    "-v",                    # Verbose output
    "--tb=short",           # Short traceback format
    "--strict-markers",     # Strict marker handling
    "-ra",                  # Show short test summary for all outcomes
]

# Add current directory to Python path for imports
pythonpath = ["."]

# Configure test markers
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Minimum version requirement
minversion = "6.0"
